Patch contributed by Anmol Khirbat.
Here is a small improvement.  When `Send Link' is invoked from a
browser tab or with only one article selected in the Articles tab, the
title of the article will be inserted into the subject of the email.
If multiple links are selected, each title will precede the
corresponding link in the body of the email instead.

This also fixes a small problem that I ran into.  Links with an
ampersand character ended up truncated in the body of the email.