Fix bug where deleting a folder in unified view mode causes Vienna to fail.
In the database code that removes the folder, a notification is posted before
the folder deletion is committed to the database. This notification is fired
to the folder tree which changes the selection to the next folder in the tree
and then fires another notification which causes that folder contents to be
refreshed in the article view. Only after that are the changes committed to 
the database.

However in unified view mode, there's a database request in the middle of the
refresh to mark all articles in the folder as read. The other views don't do
anything like that. This results in another database transaction being started
while the previous one is still in progress which is a no-no. In debug builds,
Vienna throws an assert in the debugger. In retail build, it basically just
fails and leaves the whole state incomplete.

This is dangerous. Deleting the folder should commit before firing the
notification. This fix moves the notifications outside of the transaction so
the notification handlers can safely access the database if they need to do so.