Use @synchronized() to keep database coherence, and harmonize its use
Note : Indeed, @synchronized() is known to be relatively slow. But it is Leopard compatible.

And one have to balance "performance hit" versus safety.  With Grand Central Dispatch & blocks, the behavior with an exception is undefined, since the execution of the block could happen on any thread. Blocks must catch and suppress within them any language-level exceptions thrown. Other errors that occur during the execution of the block should similarly be handled by the block itself.