Improved UnifiedDisplayView
If we detect a dimension problem, we have to wait a while until the different dimensions get coherent. This also allows us to remove the horrible hack that was needed with initial app launch : there isn't cell size problem with the standard style (Tyger graphite) and launch with a single tab.
Simplified and more efficient didFailLoadWithError WebView delegate method.