Improve accessibility of feed list items
This is a bit of a hack as we are using AXDescription for this. Overriding
AXValue is not viable without much more code because all dependent parameterized
attributes (AXAttributedStringForRange etc.) would need to be adjusted as well.

Cleaner solution can IMHO be implemented only after transitioning to a
view-based outline and have each element (error image, progress image, count
label, feed name, folder image) available as a subview of the cell’s view and
provide its own part of accessibility information. (E.g. as in Mail's mesage
list). For now, this quite works :-)