Further tweaks to User-Agent
Follow-up on issue #215 :
- default user agent will now be similar to : `Mozilla/5.0 Vienna/3.0.0`
- browser panes' agent will be similar to : `Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_2) AppleWebKit/537.74.9 (KHTML, like Gecko) Vienna/3.0.0 Safari/537.75.14`

Therefore, the browser panes User-Agent is more similar to Apple Safari's. This way, sites like iMore.com will not nag us anymore that we are using an outdated version of Safari.
Note that the first part of the browser pane User-Agent (which quotes Mozilla/5.0, OS X version and WebKit version) is automatically added by WebKit.