Moved viewDidLoad code into viewWillAppear
Preferences were not initialising on OS X < 10.10 due to the
initialisation code being placed in viewDidLoad which wasn't being
called after the transition to MASPreferences.

MASPreferences calls viewWillAppear so the initialisation code has been
moved into that. NSViewController has viewWillAppear in 10.10 and so a
check has been added before calling it on super.