Show acknowledgements in "About" panel
The "About" panel shows the contents of a Credits.rtf file, if
present. Acknowledgements.rtf is renamed into Credits.rtf and
added as a bundle resource to take advantage of this. The build
script Acknowledgements.sh is no longer necessary and the menu
item in the Help menu is removed as well.

The copyright string is reformulated and contains no more
translatable information. All the InfoPlist.strings file are
also no longer necessary.

The CFBundleGetInfoString key in Info.plist is removed, as it
has been obsolete since Snow Leopard.