Replace ToolbarItem and ToolbarButton classes
The previous implementation was based completely on images. This new implementation uses NSButton. As a result, the custom code is no longer required. Custom behaviour, such as generating smaller or grayed-out images is no longer needed either.

In most instances, it is enough to refer to NSToolbarItem instead of a subclass.