Remove FoldersTree object from MainMenu.xib
FoldersTree is now created in AppController instead of IB. FoldersTree's outlet to the outline view is replaced with a property. The outline view itself is now an outlet on AppController, which AppController will pass to FoldersTree. The outline view's outlets are also set by AppController.

The initialization happens at the -applicationDidFinishLaunching: stage in AppController. This leaves -awakeFromNib with AppController-specific setup.