Fix for passwords containing special characters
Fix issues #1203 and #1201 : we had OpenReader authentication failure
with passwords containing characters like ‘&’ ‘!’ or ‘^’.
We make the `-setPostValue:forKey` method able to safely handle these
characters by use of an algorithm similar to the one used by our
percentEscape function.

Notes :
- function CFURLCreateStringByAddingPercentEscapes: and method
stringByAddingPercentEscapesUsingEncoding: are deprecated
- using the `URLQueryAllowedCharacterSet` causes problem as this will
not escape ‘&’ or ‘+’. So we build our own character set.
- we don’t call alphanumericCharacterSet because of the enormous size of
the charset it returns.