Use CS-ID.xcconfig to sign Vienna on Deployment
As this is an open source project, the goal is to keep all informations
related to a developer or a team outside of the public repository. They
have to be set in the private CS-ID.xcconfig file.
CODE_SIGN_STYLE is Automatic for Development, Manual for Deployment.

Note the removal of some CODE_SIGN_IDENTITY lines in project.pbxproj :
this allows the value set in CS-ID.xcconfig to prevail.