Fix concurrency issues on first authentication
Fix broken dependency management and make getTokenForRequest: behave
synchronously, very similarly to addClientTokenToRequest:

Operations like 'Mark read' or 'Mark all articles read' were not working
when they were the first requested, before any authentication attempt.

The main issue with the previous logic was that any attempt to modify an
NSMutableURLRequest after a NSURLSessionTask had been prepared with it
is ineffective, even if the task is not yet in an active state.
So we were still getting some 'missing a T token' answers from the
OpenReader server, despite all our efforts.