Replace IOKit calls with NSBackgroundActivityScheduler
This API simplifies the code for the background refresh and avoids IOKit.

The system has some discretion when the handler is executed and may defer the task if the system is under load. The task priority is configured to use "utility" priority (which is between user-initiated and the default background priority).

It no longer forces a refresh upon wakeup.