Replace a localized key
The Unread localized key is shown in the window's title bar and is capitalized, even though it does not have to. There is an existing key that can be reused for this purpose. In addition, that existing key supports plural localization.

Also replaces the appName property on AppController with an NSRunningApplication property to provide the same functionality.