Fix use of the same html file by adjacent cells
When user used new browser with Unified layout, we had adjacent cells
displaying the same data.
This was caused by simultaneous access to the same "article.html" file
in cache from different calls to tableView:viewForTableColumn:row:

To fix this, we get rid of the file as soon as we don't need it anymore.
We also call setArticles: as late as possible, to minimize wait times.