Improve handling of adjacent cells (Unified layout)
Give each cell its own html file.
The files get deleted after the load was completed, when a new article
is loaded, or the article view gets deallocated.

There might be some files leaking, but the whole directory gets cleared
on the start of Vienna, so that should not result in too many issues.

Unrelated: improves Article object model.

Co-authored-by: Tassilo Karge <tassilo.karge@web.de>