Move and partially replace percentEscape() helper function
The percentEscape() helper function is used in two places: (1) for creating a `mailto:` URL and (2) for percent-encoding URLs more generally in the OpenReader class. For the first purpose, a predefined NSCharacterSet can be used instead, since `mailto:` parameters are simply URL query components. For the second purpose, the helper function is kept, but moved to OpenReader instead.