Migrate Sparkle to version 2.0 beta 1
Sparkle 2 uses XPC services, of which Vienna only needs the InstallerLauncher service. A prebuilt XPC bundle is provided in the SPM artifact, located under a SourcePackages directory in Vienna's derived-data path. The derived-data path or the SourcePackages directory cannot be referenced directly (i.e. there are no environment variables for these), so the path is referenced by going to BUILT_PRODUCTS_DIR instead and then two directories further up the directory hierarchy.

Sparkle 2 also uses a user-owned SPUStandardUpdaterController instance instead of a shared SUUpdater object. Since Vienna uses the Sparkle updater object in the main menu as well as the preferences window, it has to be used twice (s. https://sparkle-project.org/documentation/preferences-ui/). The preferences UI is controlled via Cocoa Bindings.

The Sparkle-setup.sh script is moved from a build action to a post-build action in the shared Vienna scheme, since Xcode does not guarantee that the Info.plist file is not copied over from source and preprocessed after the script is run (the result of which is that the update URL is not added). The workaround for the Xcode bug is no longer needed since Xcode 11.5.