Fix management of article's backtrack queue
Issue seen : backtracking (goBack:) which makes us switch to another
folder prevents us to return later (with goForward:) to the original
folder.

For better sequence control, we simplify reloadArrayOfArticles.
This reverses commit d35406a52e67a0839be01824508a62a41d5e6a69 which
forced fetching of articles to run on a background thread.
However, the I/O intensive -articlesWithFilter: is invoked through
dispatch_sync, so Grand Central Dispatch is allowed to run it on a
separate thread if it considers that useful.