Add printing support to the new browser
This uses private WebKit APIs on systems before macOS 11.

The printPage method of the Tab protocol is replaced by printDocument(sender)
for consistency.

When the new browser is used, the "Page Setup…" menu item in the File menu
is redundant, because the same options are available in the print panel.
Page setup is typically used by NSDocument-based apps, which Vienna is not.