Fix status text in window's status bar
Adds an `initial` key-value observing option to the window controller's observers. This causes the closures to be fired immediately, which is necessary to update the UI to reflect the latest status text. Also moves the registration/de-registration of the observers to the `-windowDidChangeOcclusionState:` delegate callback, so that observing only happens when the window is visible.