Fix toolbar search field behaviour
`sendsActionOnEndEditing` controls when a search starts. With this commit, searching is only started when the user presses the enter key. Prior to this commit, a search started when the user pressed the enter key or switched away from the search field. This makes the behaviour of the search filters more predictable, since the act of switching the filter won't start a search.

`sendsSearchStringImmediately` is unneeded, because that property only has an effect when `sendsWholeSearchString` is set to `false`.