Remove hasComments property from Article and the corresponding constants
From the git history is does not appear that Vienna has ever used this field. The database also does not have a "comments_flag" column. This field was probably added to support the "comments" element of RSS 2, but that element is supposed to be a URL to a website that has comments about the entry, not the comments themselves. Atom and JSON Feed do not support this at all.