Address several drawing issues with Xcode 15
Views no clip to their bounds anymore, which allows views to draw outside of their bounds.

In the browser UI, the RSS, cancel and reload buttons are hidden by setting layout constraints to zero (`updateAddressBarButtons()` in BrowserTab+Interface.swift). This causes the button icons to still be visible. This is addressed by enabling the `clipsToBounds` property in Interface Builder, which should be fully backwards compatible.

For the EnclosureView, the drawing is restricted to the view's `bounds` instead of `dirtyRect`.