Relax auto layout constraints inside EnclosureView
As the file image is already specified with a (default) standard size,
we can remove the superfluous size constraints on it.
Set vertical hugging priority to the value suggested by Xcode after
removal of the constraints.

Fixes issue #1618 : changing the height of the article view while
viewing an article without enclosure made the enclosure view to not
appear anymore whenever it should.