Refactor PluginManager to use NSBundle-backed plug-in objects
This adds a new Plugin class that serves as a container for NSBundle. Each plug-in can be loaded as an NSBundle instance. NSBundle creates an NSDictionary instance of the Info.plist file as well as easy access to bundled resources, like images and scripts. The Plugin class also provides read-only properties for all the Info.plist values to replace the dictionary-based plug-in store that Vienna previously had.