Localise plug-in bundles
NSBundle supports Info.plist localisation by referencing InfoPlist.strings files in .lproj subdirectories. The keys in an Info.plist file are technically all localisable. The InfoPlist.strings file in the en.lproj directory serves as the base localisation and contains the keys that should be localised.

At runtime, these localised strings are accessed as follows: when using the ‑objectForInfoDictionaryKey: method on NSBundle, the bundle will return the value from the InfoPlist.strings file if one exists for the user's chosen localisation. If not, it will take the value from the Info.plist file instead.

Because search plug-ins do not use the tooltip key, that key is removed from the Info.plist.