Revert recovery procedure, prevent sibling loss
I experienced that the recovery procedure led to an incoherent tree,
so the best workaround remains switching back to alphabetical auto sort
by reverting commit e3751a9. However, reselect user's choice after that.

As mail exchanges with user @liamgreenwood around issue #1768 hinted
that "loss" of a sibling is the likely origin of -loadTree:rootNode's
failure, try to find it and reattach it as soon as it is reported missing.