Handle command-click on links in browser tabs
Make command-click on a link in a browser tab to open a new tab, like in
other major web browsers.
Also have option-click in a browser tab behave the same way as an option
-click in the primary tab : open the link in the browser which is not
set as the default in General settings (either it is the system's default
browser or Vienna's internal browser).

Solves request expressed in discussion #1740