Stop recurrent unsuccessful attempts to get tokens
When an authentication token is required, application's main thread
has to wait for responses from the Open Reader server.
So when we perform "Refresh All Subscriptions" while the requests to the
server are timing out, we risk having the application UI freezing for a
significant time.
Workaround : drop following token requests for this batch of requests.