Improve handling and storage of font selection
The `articleListFont` (i.e. font name) and `articleListFontSize` settings are replaced by the `articleListFont` (NSFont instance) setting. Selecting a font in the settings window stores the NSFont instance (typeface, style and size) as-is in user defaults, instead of another instance that is based only on the name and size of the chosen font. The settings window and article list will work with the stored font too. Retrieving the stored font is more resistant to invalid data; Vienna's default font will be used if the stored font is invalid.