Make predicate editor more robust to spacing changes
If I moved a feed (or a folder) into a group folder or out from it, the
editor got confused when I attempted to edit a preexisting smart folder
involving this feed/folder, because the number of indentation spaces had
changed.
More sophisticated solutions would be possible, but keep it simple and
stupid for now.

Partial solution to issue #1904 (fragility of folder selection in smart
folders)