Remove conflicting selection colours in ArticleListView
These two colours were applied to selected and clicked cells, but they
had no effect on selected cells and they resulted in unreadable text on
clicked cells when not using dark mode.

When a cell is selected, the cell's text colour is overridden to match
the highlight colour.

When a cell is clicked (right click), the cell is drawn with a border
using the highlight colour. The text colour does not have be changed.