Replace NSToolbarItem subclass with NSMenuToolbarItem
In macOS 15, the PopUpButtonToolbarItem subclass is not compatible with
the text-only mode of NSToolbar. NSMenuToolbarItem (macOS 10.15+) works
and simplifies the code, but is not available in Interface Builder. The
pop-up button toolbar items are now set up in code to take advantage of
NSMenuToolbarItem.

The PopUpButtonToolbarItem class is updated accordingly. Text-only mode
should still work on macOS 10.13 and 10.14 for which this class will be
used.