Remove BrowserTabWithLegacyAddressBar and the relevant image assets
This file is now only used for macOS 10.13, which can use the BrowserTab
file. The only side effect is that the address bar cannot be put on top
of the web view, since content insets (see: _topContentInset) are not
available until macOS 10.14. Instead, the address bar is placed above
the web view, even if this means that there will not be a vibrancy
effect (i.e. a minor visual issue).